+++ CHANGELOG.md
@@ -54,6 +54,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Provided the `lang` attribute to `<html>` `iStaticPage`
 * Fixed a bug when providing `href` to `dataProvider` `bButton`
 * Fixed a bug when mutations of the nested path can't be caught `super/i-block/modules/sync`
+* Fixed watching with parameters `{immediate: true, collapse: true}` `core/component/watch`
 
 ## v3.0.0-rc.210 (2021-07-07)